+++ include/rnc/assembly_textbin_set_tfrag_dists_types.h
@@ -4,9 +4,9 @@
 #include "types.h"
 
 struct M2c_D_00160EA0 {
-    s32 unk0;
-    s32 unk4;
-    s32 unk8;
+    f32 unk0;
+    f32 unk4;
+    f32 unk8;
 };
 
 struct M2c_D_00160EB0 {
@@ -17,18 +17,18 @@ struct M2c_D_00160EB0 {
 
 struct M2c_D_001DE7F0 {
     s32 unk0;
-    s32 unk4;
+    f32 unk4;
     u8 pad_8[0x4];
-    s32 unkC;
-    s32 unk10;
-    s32 unk14;
+    f32 unkC;
+    f32 unk10;
+    f32 unk14;
     u8 pad_18[0x4];
-    s32 unk1C;
-    s32 unk20;
-    s32 unk24;
+    f32 unk1C;
+    f32 unk20;
+    f32 unk24;
     u8 pad_28[0x8];
-    s32 unk30;
-    s32 unk34;
+    f32 unk30;
+    f32 unk34;
 };
 
 #endif /* RNC_ASSEMBLY_TEXTBIN_SET_TFRAG_DISTS_TYPES_H */
